separate linear from module map
