+++ python/metatensor-learn/metatensor/learn/nn/__init__.py
@@ -1,4 +1,5 @@
-from .module_map import ModuleMap, Linear  # noqa
+from .module_map import ModuleMap # noqa
+from .linear import Linear # noqa
 
 
 ___all__ = ["ModuleMap", "Linear"]
