+++ python/metatensor-torch/tests/learn/module_map.py
@@ -3,7 +3,7 @@ import torch
 from torch.nn import Module, Sigmoid
 
 from metatensor.torch import allclose_raise
-from metatensor.torch.learn.nn import Linear, ModuleMap
+from metatensor.torch.learn.nn import ModuleMap
 
 from .utils import TORCH_KWARGS, random_single_block_no_components_tensor_map
 
@@ -17,7 +17,6 @@ class MockModule(Module):
 
     def forward(self, input: torch.Tensor) -> torch.Tensor:
         return self._last_layer(self._activation(self._linear(input)))
-
 
 class TestModuleMap:
     @pytest.fixture(autouse=True)
@@ -74,76 +73,6 @@ class TestModuleMap:
             random_single_block_no_components_tensor_map(True, True),
         ],
     )
-    def test_linear_module_init(self, tensor):
-        # testing initialization by non sequence arguments
-        tensor_module_init_nonseq = Linear(
-            in_keys=tensor.keys,
-            in_features=[2],
-            out_features=[2],
-            bias=[False],
-            out_properties=[tensor[0].properties],
-        )
-        # testing initialization by sequence arguments
-        tensor_module_init_seq = Linear(
-            in_keys=tensor.keys,
-            in_features=2,
-            out_features=2,
-            bias=False,
-            out_properties=tensor[0].properties,
-        )
-        for i in range(len(tensor_module_init_seq)):
-            assert (
-                tensor_module_init_seq[i].in_features
-                == tensor_module_init_nonseq[i].in_features
-            ), (
-                "in_features differ when using sequential and non sequential input for"
-                " initialization"
-            )
-            assert (
-                tensor_module_init_seq[i].out_features
-                == tensor_module_init_nonseq[i].out_features
-            ), (
-                "out_features differ when using sequential and non sequential input for"
-                " initialization"
-            )
-            assert (
-                tensor_module_init_seq[i].bias == tensor_module_init_nonseq[i].bias
-            ), (
-                "bias differ when using sequential and non sequential input for"
-                " initialization"
-            )
-
-        tensor_module = tensor_module_init_nonseq
-
-        with torch.no_grad():
-            out_tensor = tensor_module(tensor)
-
-        for i, item in enumerate(tensor.items()):
-            key, block = item
-            module = tensor_module[i]
-            assert (
-                tensor_module.get_module(key) is module
-            ), "modules should be initialized in the same order as keys"
-
-            with torch.no_grad():
-                ref_values = module(block.values)
-            out_block = out_tensor[i]
-            assert torch.allclose(ref_values, out_block.values)
-            assert block.properties == out_block.properties
-
-            for parameter, gradient in block.gradients():
-                with torch.no_grad():
-                    ref_gradient_values = module(gradient.values)
-                out_gradient = out_block.gradient(parameter)
-                assert torch.allclose(ref_gradient_values, out_gradient.values)
-                assert gradient.properties == out_gradient.properties
-
-    @pytest.mark.parametrize(
-        "tensor",
-        [
-            random_single_block_no_components_tensor_map(True, True),
-        ],
-    )
     def test_torchscript_module_tensor(self, tensor):
         modules = []
         for key in tensor.keys:
@@ -153,26 +82,6 @@ class TestModuleMap:
                 )
             )
         tensor_module = ModuleMap(tensor.keys, modules)
-        ref_tensor = tensor_module(tensor)
-
-        tensor_module_script = torch.jit.script(tensor_module)
-        out_tensor = tensor_module_script(tensor)
-
-        allclose_raise(ref_tensor, out_tensor)
-
-        # tests if member functions work that do not appear in forward
-        tensor_module_script.get_module(tensor.keys[0])
-
-    @pytest.mark.parametrize(
-        "tensor",
-        [
-            random_single_block_no_components_tensor_map(True, True),
-        ],
-    )
-    def test_torchscript_linear_module(self, tensor):
-        tensor_module = Linear(
-            tensor.keys, in_features=len(tensor[0].properties), out_features=5
-        )
         ref_tensor = tensor_module(tensor)
 
         tensor_module_script = torch.jit.script(tensor_module)
